ENH: Configuration system for Py-ART
Various Py-ART settings can now be customized using a Py-ART configuration
file.  The location of this file should be defined in the `PYART_CONFIG`
environmental variable.  A sample configuration file with instructions for
how to customize it is contained in the Py-ART source code in the file
`default_config.py`.

A short list of features introduced by this enhancement:

* Customized mapping of field names for a given file type.
* All read functions have a field_mapping, additional_metadata, and
  exclude_fields parameters.
* A load_config function for loading configuration files on the fly.
* Documentation and unit tests for new features.
* Field names for correction and mapping algorithms are no longer hard coded
  and can set in the configuration file.
* The fillvalue can be set in the configuration file.