ENH: 4DD routine
Adding the capability to use a previous velocity volume as an initial condition
to the dealiasing routine. Note that this enhancement was how the original 4DD
algorithm was intended to be implemented, as it should provide better results.
This commit attempts to address some of the issues raised in issue #119.