ENH: Grid object
Adding the ability to exclude fields during the creation
of a Grid object, as well as the ability to add a new field
with the 'add_field' attribute similar to the Radar object.
Furthermore, I am adding some extra axes and dimensions to check
for when creating a Grid object. These axes and dimensions are
compatible with older versions of MMCG files.