MOD: GridMapDisplay
Making the necessary changes such that the Basemap plots work as
expected. One of the changes required was to make sure the grid origin
latitude and longitude were not NumPy arrays (of size 1) but scalar
values. This was previously affecting the output of pyproj.Proj() when
using these as lat_0 and lon_0.