ADD: echo classification
Adding echo classification module, including a Fortran routine for the
method of Steiner et al. (1995). The echo classification module should
be general and host any number of methods (e.g. Steiner method,
Biggerstaff method, etc). This commit does not include the necessary
object and/or pyf files required to actually run the code, as I figured
@jjhelmus would be best at handling this. I assume some modification to
the setup.py file in pyart.retrieve is in order...