ENH: Grid quality control
Adding a field 'despeckling' method to the Grid object. This method can
be used to remove outliers from gridded fields. The heavy lifting of
this routine is done in Fortran. This routine defines a 3-D box filter,
counts the total number of points within this box as well as how many of
these points are 'noise', and if this ratio is too high, flags the grid
point as noise.