ADD: echo class wrapper
Adding f2py wrapper to the Fortran Steiner echo classification routine.
According to the SciPy docs, the distutils extensions in NumPy can be
used in setup.py to define an extension module entirely in terms of this
pyf interface file, and is the best way to distribute f2py extension
modules.