ENH: objective analysis
I am beginning to update and/or enhance the grid_mapper module in
Py-ART. As it stands now, the Cressman and Barnes filters (weights) are
poorly defined compared to the literature, e.g. Trapp and Doswell
(2000). In the literature, the Cressman weight is a function of the
radius of influence, while the Barnes weight is a function of the
smoothing parameter. However in the grid_mapper module, the user
currently only has control over the radius of influence, which will
likely confuse anyone who wants to use the Barnes weight.

Section 2b of Trapp and Doswell (2000) steps the reader through how to
"objectively" determine an appropriate smoothness parameter for the
Barnes weight for dealing with data of variable spacing, like that from
a scanning radar. However as the grid_mapper module stands now this is
not necessarily intuitive since there is only a radius of influence
parameter option.