ADD: correct Barnes weight
Adding the ability for the user to specify the free parameters for the
Barnes distance-dependent weighting function. In Trapp and Doswell
(2000), the smoothing parameter (kappa) is the only free parameter,
however it is rewritten as a function of the nondimensional smoothing
parameter (kappa star) and the data spacing, which for a scanning radar
should be the maximum data spacing within the analysis domain.

As of now the "modified" Barnes weight, like that in section 2c of Trapp
and Doswell (2000) is not available. The smoothing parameter for the
modified Barnes weight is allowed to change with the local data spacing,
similar to the variable radius of influence for the Cressman weight.