ENH: radar location within analysis domain
Swapping the pyproj package in for the corner_to_point function, which
should give more accurate results for the Cartesian position of a radar
within the analysis domain, provided a reasonable projection argument is
given. The default is 'lcc', which is the Lambert Conformal Conic
projection.

A simple test with the SGP C-SAPR showed that the difference in the x
and y position of the radar within the analysis domain differed between
the two techniques by approximately 7 m and 140 m, respectively.