FIX: bug fix and minor enhancements
Fixing a minor bug in the Grid object as to how it handles fields with
incorrect shapes. Also adding the capability for the user to exclude fields
during a read and an add_field method to the Grid object, both of these
similar to their counterparts in the Radar object.