BUG: Treat radar coordinates as gate centers in plots
The range, azimuth, elevation coordinates in radar files typically specify
the center of the gate, when plotting Py-ART uses these coordinates to
interpolate and extrapolate the gates edges which are then used to plot
the data.

Prior to this the centers were used to plot which resulted in the plots
missing the last range gate and the last ray.