ENH: Display object now respect antenna_transition
The plot method of the various Display classes (RadarDisplay, RadarMapDisplay,
etc) will not filter out rays which are marked as in-transition by the
antenna_transition attribute of the radar.  The behavior can be turned off by
setting the filter_transitions parameter to False.  When no antenna transition
information is available (antenna_transition is None) no filtering is performed
regardless of the value of the filter_transitions parameter.