ENH: add inclusive kwarg to filters
Adding an 'inclusive' keyword to the relevant filter methods (e.g.,
exclude_below). This gives the user more control over the specifics and
reduces ambiguity, for example, it's ambiguous as to whether the actual
minimum and maximum values specified in exclude_inside (include_inside)
will be excluded (included) or not unless the user dives into the actual
code. One final point is that the default setting of the inclusive kwarg
is consistent with what @jjhelmus originally intended with the filter
methods.