BUG: Dealias gates with vel outside of the Nyquist
Gates with velocities outside of the Nyquist velocities are unwrapped when
the dealias_region_based function is used.  Previously these gates would be
ignored by the algorithm unless the interval_limits parameter was specified to
include these values.

A warning is issued when gates with velocities outside the Nyquist are
detected as this could be an indication of a more serious problem.

closes #379