ENH: Decorator for checking for dependencies at run time
The requires function in pyart.pkg_util.requires_decorator can be used to
decorate functions or method which depend on an optional depency.  The function
or class containing the method can then be imported ragardless of if the
dependency is met.  If the function or method in question is called and the
optional depency is not met a MissingOptionalDepency will be raises with a help
message indicating that the depency is not installed and should be.