ENH: Radar.get_gate_x_y_z method
Add a get_gate_x_y_z method to the Radar class for obtaining the gate center
or edge distances from the radar.

Remove the gate_edge_{x, y, z} attributes from the Radar class. For radar
volumes with multiple sweeps these attributes did not correctly give the gate
edge locations. The correct edge location must be calculated on a
sweep-by-sweep basic and can be obtained from the get_gate_x_y_z method.
Since this attribute was only recently added to the class and was never in a
formal release, no deprecation is needed.