MAINT: Find gate locations from geographic coords
When gridding radar data using the map_gates_to_grid function find the location
of the radar gates using the geographic coordinate gates calculated using map
projection set in the Radar instance.

Use the geographic_to_cartesian_aeqd function to find the displacement of these
gates and the radars themselves to the grid origin.  Eventually the projection
used in the grid should be used to find these displacements.