BUG: map_to_grid gate locations from geographic
When gridding radar data using the map_to_grid function find the location
of the radar gates using the geographic coordinate gates calculated using the
map projection set in the Radar instance.

Use the geographic_to_cartesian_aeqd function to find the displacement of the
gates and the radars themselves to the grid origin.  Eventually the projection
used in the grid should be used to find these displacements.