DOC: Building of examples opt-in
The default behavior when build the Py-ART documentation is to not build the
documentation for the examples. This can be changed by setting a
BUILD_PYART_EXAMPLES environment variables.  Alternatively one of the three
rebuild_ scripts in the docs folder can be used:

 * rebuild_docs.sh : Rebuild standard docs, examples built only if variable set.
 * rebuild_example.sh : Rebuild the example (and the other docs).
 * rebuild_full_docs.sh : Rebuild the full docs including the examples.

If the examples have never been built a single warning will be raised about the
lack of examples.  This can be ignored.