ENH: near range BC check
The main change in this commit has to do with adding a check for near
range boundary condition outliers. Ground clutter close to the radar can
introduce spurious differential phase values that, if left unchecked,
can severely affect the quality of propagation differential phase and
thus specific differential phase. The check that I've added can also be
thought of as a way to estimate the system phase offset assuming the
nearest range gates are mostly within the first few range gates. The key
assumption with this check is that the "real" near range BC signal is
the main signal and dwarfs the ground clutter signal.