BUG: ray has all contiguous gates
This bug was found during the writing of some unit tests, so score
another point for unit tests. Basically if the ray has no masked data
then only one slice object would be available for that ray and a
TypeError would be raised trying to loop over the slice object.

I also moved the change_outliers keyward to **kwargs so that it could be
passed directly from kdp_maesaka() if necessary. This may be bad Python
though...