ENH: kdp_proc unit tests
Adding unit tests for the kdp_proc module. So far this includes two
tests, one to make sure retrieved KDP approaches the first guess if all
gates are excluded in the input GateFilter, and the second to make sure
retrieved KDP approaches a constant value for a linear PSIDP profile.
This latter test can be enhanced or a new test altogether should be
written to make sure that the constant value equals the slope of the
linear PSIDP profile.