MAINT: unit tests for kdp_proc
The unit tests for kdp_proc module now include the test for KDP
converging to the slope (half the slope actually) of a linear
differential phase profile. I removed the docstring for each test
function to be consistent with the other unit tests and with how their
results are reported in AppVeyor.

I also changed the default first_guess value from 0.001 to 0.01 since I
noticed during testing with real data that a value closer to 1 rather
than 0 is likely more beneficial for avoiding a precision loss warning
during minimization.