BUG: Raise an error for too much sounding data
Raise a ValueError when more than 999 sounding heights are provided to the
dealias_fourdd function.  This prevents a segmentation fault from occuring in
the C code of the FourDD algorithm.

Modifying the C routine where this limit is imposed to something larged (~5000)
would be reasonable but this change should wait until after the next release
in order for this error to be caught.

closes #571