ENH: Added local install option for radx
Added a keyword argument to pyart.aux_io.read_radx radx_dir which allows for the
user to define a local (non-root) install of RadX. This should be very handy for
users who just want to drop in the pre-compiled radx binaries. Tested on OSX
sierra with CPOL RAPIC files.