ADD: Version Checking
While I was combing through the ICD I realized there's a version number for each product included in the product description block that should be a good way to check that the product hasn't changed to a new version that is not supported by Pyart. If it has it issues a warning which hopefully should cause someone to take a look at the most recent ICD and verify there are no major changes.

I did not add version checking for the metadata parts that come before each product's data. I think there's a potential for these to change from build to build in a way that could break something, but also a greater potential to result in annoying warnings for no reason.

The good news is version numbers don't seem to change often. I compared Build 18 to Build 13, which is what this was originally based on, and found that they only increased the version number for products 165 and 172. Both were relatively minor. I checked that the new files uploaded successfully and they did. I couldn't find the parameters the ICD mentions that have changed for 172; I assume Pyart doesn't parse these. For 165 I added the Large Hail and Giant Hail classifications for user interpretation of the classification codes.