ADD: Utility to subset a radar volume for the column above a location (#1113)
* ADD: Utility to subset a radar volume for the column above a location
given the lat/lon of the location. Return is a dictionary containing all radar object fields for the column above the location, as well as, x,y,z gates for the column. Tested with ppi (nexrad) and rhi (csapr) scans

* MOD: first corrections to initial PR for columnsect.py
Still working on transfering output to an xarry object

* ADD: Unit Tests for sphere_distance, for_azimuth, and get_field_location functions within
columnsect.py utility

* MOD: Added the requested changes from reviewers. Main modifications include changing the output of
get_field_location form a dictionary to xarray Datatset.