Minor adjustments to convective-stratiform algorith, documentation, and example (#1434)
* Fix spacing in documentation

* Fixing overlapping plots in example

* Fix minor issue causing arrays to be incorrectly masked

This fix prevents the final convective stratiform array from being masked by the background field. Due to the nature of the `calc_thres` (minimum number of points needed to calculate the background average), the background field does not compute on the edges of the echo. Instead, we want the convective stratiform field to cover the same region as the original reflectivity field.

* Fix lint error

* Fix linting error

* Fixing plot overlap in example