ENH: Add 'zdist_factor' input parameter to 'map_gates_to_grid'. (#1509)
* FIX: change deprecated np.str in arm_vpt to str

* Update default wind_size in calculate_velocity_texture from 4 to 3
Using an odd number prevents single index offsets in placement in
case of an even size

* ADD: optional rectangular window dims for velocity texture analysis.
Also, modify the default window size to 3 instead of the previous
value of 4 (even number resulting in a potential offset).

* FIX: nyquist velocity dtype in radar obj loaded with read_kazr

* pep8

* line reformatting per black

* TST: unit tests for rectangular texture window

* STY: linting (black)

* FIX: Add in linting fixes

* FIX: metadata in core.radar.get_elevation

* FIX: outdated text in masking_data_with_gatefilters.ipynb

* ADD: method to automatically determine sweep number and sweep start and end indices

* add comment to 'determine_sweeps' method

* FIX: indexing bug in 'determine_sweeps' method

* ADD: unit tests for 'determine_sweeps'

* linting

* FIX: nsweeps wasn't updated when calling 'determine_sweeps'; tests were updated accordingly

* line reformatting per black

* FIX: update multiple radar object attributes to ensure its full utilization with Py-ART after 'determine_sweeps' is called

* line reformatting per black

* linting

* ENH: add antenna_transition to the testing module's sample objects

* ENH: update unit tests to reflect antenna_transition allocation in sample_objects

* Update pyart/util/radar_utils.py

Co-authored-by: Max Grover <mgroverwx@gmail.com>

* Update pyart/util/radar_utils.py

Co-authored-by: Max Grover <mgroverwx@gmail.com>

* ENH: add 'zdist_factor' input parameter to 'map_gates_to_grid'.
This parameter scales the distance in the z-dimension when calculating
weights upon gridding. It is a semi-equivalent to the scaling factors in
the RoI class methods. Example for using this parameter would be setting
a 'zdist_factor' value of 0.0 combined with an h_factor=0.0 (if calling
DistBeamRoI) or z_factor=0.0 (if calling DistRoI), resulting in the
exclusion of the z dimension in gridding weighting, which could serve as
a potential solution for gridding a single PPI sweep from a single radar.
In that case, the distance in the weighting function effectively serves as
the distance of a given point from the PPI sweep's cone.

* DEL: remove a redundant (liekly residual) input parameter (toa) from map_gates_to_grid

* MNT: set a consistent nomenclature for the distance squared parameter.
In parts of the 'map_grid' method it was referred to as 'dist' whereas
in another as 'dist2' (parameter calculation lines differ in style but
identical in value), resulting in an unnecessary scalar allocation,
and a confusing nomenclature.

* DOC: Add example for gridding data

* Update pyart/map/gates_to_grid.py

Co-authored-by: Max Grover <mgroverwx@gmail.com>

---------

Co-authored-by: mgrover1 <mgroverwx@gmail.com>